Rename to ExtensionTemplate
