+++ lib/buildkite/builder/extension_template.rb
@@ -1,0 +1,13 @@
+module Buildkite
+  module Builder
+    class ExtensionTemplate
+      attr_reader :extension_class, :name, :block
+
+      def initialize(extension_class, name, block)
+        @extension_class = extension_class
+        @name = name
+        @block = block
+      end
+    end
+  end
+end
